Separated usage list transformation into its own class
