+++ databricksusagereport/graph/graph.py
@@ -14,20 +14,10 @@ class Graph:
         pass
 
     @staticmethod
-    def create(usage_list):
-        usage_list_transformed = dict()
-
-        for usage in usage_list:
-            if usage["name"] not in usage_list_transformed.keys():
-                usage_list_transformed[usage["name"]] = {"x": [usage["date"]],
-                                                         "y": [usage["NumWorkers"]]}
-            else:
-                usage_list_transformed[usage["name"]]["x"].append(usage["date"])
-                usage_list_transformed[usage["name"]]["y"].append(usage["NumWorkers"])
-
+    def populate_template(usage_list, title, mode):
         graph_data = ""
 
-        for key, value in usage_list_transformed.iteritems():
+        for key, value in usage_list.iteritems():
             # Sort X an Y values by X to achieve chronological order
             y = [y for (x, y) in sorted(zip(value["x"], value["y"]))]
             x = sorted(value["x"])
@@ -36,10 +26,9 @@ class Graph:
                                                        "'%s'" % "', '".join(map(str, x)),
                                                        ', '.join(map(str, y)),
                                                        key,
-                                                       "lines+markers")
+                                                       mode)
 
-        graph_data_list = ', '.join(map(str, usage_list_transformed.keys()))
-        title = "Cluster Worker Usage"
+        graph_data_list = ', '.join(map(str, usage_list.keys()))
 
         filled_template = Graph.databricks_graph_data_template % (graph_data,
                                                                   graph_data_list,
